requirements.txt
pandas
nltk==3.9.3
mlflow==3.10.1
pyyaml

transformers==4.45.0
accelerate==0.26.0
